tools/_shared/pyproject.toml
[project]
name = "outreach-shared"
version = "0.1.0"
description = "Cross-platform shared library for outreach tools (X, XHS)"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30",
    "google-genai>=1.0",
    "playwright>=1.49",
    "playwright-stealth>=1.0.6",
    "structlog>=24.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.9",
    "mypy>=1.14",
    "testcontainers[postgres]>=4.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
